requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
httpx
sqlalchemy
alembic
python-jose[cryptography]
bcrypt
python-multipart
orjson
psycopg[binary]
cryptography
pytest
